Fix dialog UIs. Do not use AlertDialog.title property.
